Move _math submodule to math submodule.
