+++ hcipy/fourier/chirp_z_transform.py
@@ -2,7 +2,7 @@ import numpy as np
 from scipy.fft import next_fast_len
 
 from .fourier_transform import _get_float_and_complex_dtype
-from .._math import fft as _fft_module
+from ..math._implementation import fft as _fft_module
 
 class ChirpZTransform:
     '''The Chirp Z-transform (CZT).
