+++ hcipy/math/_implementation/cpu.py
